+++ packages/svelte/src/index.ts
@@ -1,3 +1,16 @@
-export * from './use-chat';
-export * from './use-completion';
-export * from './use-assistant';
+export {
+  Chat,
+  type ChatOptions,
+  type CreateMessage,
+  type Message,
+  type UIMessage,
+} from './chat.svelte.js';
+
+export {
+  StructuredObject as Experimental_StructuredObject,
+  type Experimental_StructuredObjectOptions,
+} from './structured-object.svelte.js';
+
+export { Completion, type CompletionOptions } from './completion.svelte.js';
+
+export { createAIContext } from './context-provider.js';